Renamed and moved DeactivatedMemberCleaner
